Handle error in apk extraction

+++ app/src/main/java/in/arjsna/permissionchecker/appdetails/AppDetailsPresenterImpl.java
@@ -2,16 +2,16 @@ package in.arjsna.permissionchecker.appdetails;
 
 import android.content.Context;
 import android.os.Environment;
+import android.support.annotation.NonNull;
 import in.arjsna.permissionchecker.basemvp.BasePresenter;
 import in.arjsna.permissionchecker.datamanager.DataProvider;
 import in.arjsna.permissionchecker.di.qualifiers.ActivityContext;
 import in.arjsna.permissionchecker.models.AppDetails;
 import in.arjsna.permissionchecker.models.PermissionDetail;
-import io.reactivex.Completable;
 import io.reactivex.Single;
+import io.reactivex.SingleEmitter;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.CompositeDisposable;
-import io.reactivex.observers.DisposableCompletableObserver;
 import io.reactivex.observers.DisposableSingleObserver;
 import io.reactivex.schedulers.Schedulers;
 import java.io.File;
@@ -24,6 +24,7 @@ public class AppDetailsPresenterImpl<V extends IAppDetailsView> extends BasePres
 
   private ArrayList<PermissionDetail> permissionDetails;
   private AppDetails appDetails;
+  private boolean canReplaceExistingFile = false;
 
   @Inject public AppDetailsPresenterImpl(@ActivityContext Context context,
       CompositeDisposable compositeDisposable, DataProvider dataProvider) {
@@ -131,10 +132,19 @@ public class AppDetailsPresenterImpl<V extends IAppDetailsView> extends BasePres
   }
 
   @Override public void onPermissionDenied() {
-
+    getView().showError("APK extraction failed");
   }
 
   @Override public void onPermissionGranted() {
+    extractAPK();
+  }
+
+  @Override public void extractByReplacing() {
+    canReplaceExistingFile = true;
+    extractAPK();
+  }
+
+  private void extractAPK() {
     getCompositeDisposable().add(createDirectoryAndExtractApk().subscribeOn(Schedulers.io())
         .observeOn(AndroidSchedulers.mainThread())
         .subscribeWith(new DisposableSingleObserver<String>() {
@@ -143,6 +153,10 @@ public class AppDetailsPresenterImpl<V extends IAppDetailsView> extends BasePres
           }
 
           @Override public void onError(Throwable e) {
+            if (e instanceof IllegalStateException) {
+              getView().showFileExitsAlert();
+              return;
+            }
             getView().showError("APK extraction failed");
           }
         }));
@@ -150,18 +164,28 @@ public class AppDetailsPresenterImpl<V extends IAppDetailsView> extends BasePres
 
   private Single<String> createDirectoryAndExtractApk() {
     return Single.create(e -> {
-      File appDir =
-          new File(Environment.getExternalStorageDirectory() + "/AppPermissionsExtractedApk");
-      if (!appDir.exists()) {
-        if (!appDir.mkdir()) {
-          e.onError(new Exception("Error creating directory"));
-        }
-      }
-      // TODO: 22/11/17 check if it already exits
+      String appDirPath = createDirectory(e);
       File apkFile = new File(appDetails.publicSrcDir);
-      File destFile = new File(appDir.getPath() + "/" + appDetails.name + ".apk");
-      FileUtils.copyFile(apkFile, destFile);
-      e.onSuccess(destFile.getPath());
+      File destFile = new File(appDirPath + "/" + appDetails.name + ".apk");
+      if (!destFile.exists() || canReplaceExistingFile) {
+        FileUtils.copyFile(apkFile, destFile);
+        e.onSuccess(destFile.getPath());
+      } else {
+        e.onError(new IllegalStateException("File already exits"));
+      }
     });
+  }
+
+
+
+  @NonNull private String createDirectory(SingleEmitter<String> e) {
+    File appDir =
+        new File(Environment.getExternalStorageDirectory() + "/AppPermissionsExtractedApk");
+    if (!appDir.exists()) {
+      if (!appDir.mkdir()) {
+        e.onError(new Exception("Error creating directory"));
+      }
+    }
+    return appDir.getPath();
   }
 }
